% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_LPUE.R
\name{add_LPUE}
\alias{add_LPUE}
\title{Add a field related to Landings Per Unit of Effort (LPUE) in the dataframe derived based on MCDA}
\usage{
add_LPUE(
  data_FPI = sample_FPIq,
  spec = c("HKE", "MUT"),
  folder = "",
  name = c("hke.txt", "mut.txt"),
  field = c("grid_code", "grid_code"),
  LON = "POINT_X",
  LAT = "POINT_Y",
  sep = ","
)
}
\arguments{
\item{spec}{An object including the species that will be included in the MCDA dataframe in 3 alpha code format. For example c("HKE","MUT")}

\item{folder}{The folder path that contains the species distribution outcomes (in a text format e.g. txt, csv)}

\item{name}{An object with the name of the files. For example c("hke.txt","mut.txt").}

\item{field}{The field which is containing the predicted values of species distribution that will be used as a proxy for LPUE. For example c("grid_code","grid_code")}

\item{LON}{The field containing longitude values.The field should be in "".}

\item{LAT}{The field containing latitude values.The field should be in "".}

\item{sep}{The separator}

\item{data}{The dataframe derived by MCDA}
}
\value{
Return a dataframe including a field to be used as a proxy for LPUE and species.
}
\description{
Using the predicted species distribution to estimate a proxy of LPUE. The fields for LPUE by species are added to the dataframe derived from MCDA. The above is used for estimating landings' weight and value.
}
\examples{
library(tools4MCDA)
data(sample_FPIq)
data_with_LPUE<-add_LPUE(data_FPI=sample_FPIq,spec=c("HKE","MUT"),folder="_Your_folder_",name=c("hke.txt","mut.txt"), field=c("grid_code","grid_code"), LON="POINT_X", LAT="POINT_Y", sep="," )
}
\author{
I. Maina
}
