% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot_year}
\alias{bubble_plot_year}
\title{Create a bubble map by year}
\usage{
bubble_plot_year(
  data,
  field,
  map_sf,
  xlim = c(19, 24),
  ylim = c(36, 40),
  years,
  ncol = 3,
  classes = waiver(),
  LegendTitle,
  PlotTitle,
  limits = NULL,
  max_size = 6
)
}
\arguments{
\item{data}{a dataframe with the Activity index by port}

\item{field}{the field to be plotted. The field should be previously classified using the tools4MCDA::classify function.}

\item{map_sf}{the polygon for the base map in sf format.}

\item{xlim}{limits on x axis}

\item{ylim}{limits on y axis}

\item{years}{column including years in the data}

\item{ncol}{number of columns of the plot}

\item{classes}{number of the classes or function}

\item{LegendTitle}{The title of the legend}

\item{PlotTitle}{The title of the plot}

\item{limits}{}

\item{max_size}{maximum size of the bubble}
}
\value{
Returns a plot
}
\description{
Create a bubble map by year to plot the Activity index by port.
}
\examples{
library(tools4MCDA)
library(sf)
library(sfheaders)
library(ggplot2)
library(tidyr)
library(BAMMtools)

data("fleet_reg_year")

#Define year and country
ye<-2019
country="GRC"

# unite gear_type and vessel_length
fleet_reg_year<-unite(fleet_reg_year, col='key', c('MAIN_FISHING_GEAR', 'vessel_length'), sep='_', remove=FALSE)

#Bubble maps by year

#base map for bubble plots
poly<- map_data("world")
map_sf <- sfheaders::sf_polygon(obj = poly, x = "long", y = "lat", polygon_id = "group")
sf::st_crs(map_sf) <- 4326

# classify values for plotting
fleet_reg_year<-fleet_reg_year[fleet_reg_year[1]==country,]
i<-"GTR_VL0612"
fleet<-fleet_reg_year[fleet_reg_year$key==i,]
classes<-as.data.frame(getJenksBreaks(as.numeric(fleet$SUM_LOA_GT),7))
colnames(classes)<-"classes"
classes$key<-i
cla3<-as.numeric(classes[,1])
cla3<-cla3[-7]
fl_yy<-fleet_reg_year[fleet_reg_year$key==i,]
fl_yy$kmcl<-as.numeric(tools4MCDA::classify(data_field=as.numeric(fl_yy[,"SUM_LOA_GT"]),cuts=cla3, scores=cla3))
fl_yy$x<-as.numeric(fl_yy$LONGITUDE)
fl_yy$y<-as.numeric(fl_yy$LATITUDE)
fl_yy$SUM_LOA_GT<-as.numeric(fl_yy$SUM_LOA_GT)
fl_yy<-fl_yy[fl_yy$YEAR==ye,]

# bubble plot
bubble_plot_year(data=fl_yy,field='kmcl',map_sf=map_sf,
                xlim=c(as.numeric(min(fl_yy$x)),as.numeric(max(fl_yy$x))),ylim=c(as.numeric(min(fl_yy$y)),as.numeric(max(fl_yy$y))), years=ye,
                classes=cla3, ncol=1,
                limits=(c(0,as.numeric(max(fl_yy[,"SUM_LOA_GT"])))),PlotTitle=paste0(i,fl_yy$YEAR),
                LegendTitle=paste0(i,fl_yy$YEAR), max_size = 4)
}
\author{
I. Maina
}
