% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_FDays_v030424_RDBFIS.R
\name{check_FDays}
\alias{check_FDays}
\title{Compare fishing effort in days at sea estimated using Fishing Pressure Index (from MCDA) as a proxy and table_G on fishing effort estimations by GSA/gear/metier/quarter/fishing_tech/target_assemblage/year/country}
\usage{
check_FDays(
  data,
  Sub.region,
  Country,
  FDays,
  table_g_effort,
  gear,
  metier,
  fishing_tech,
  target_assemblage,
  year,
  quarter,
  vessel_length_cat,
  LON,
  LAT
)
}
\arguments{
\item{data}{A data frame with Fishing effort estimations in Days at Sea in fine spatial scale. The above data frame should contain the fields: LON(longitude in WGS84), LAT(latitude in WGS84), gear(values: GNS, GTR, LLS), year, vessel_length_cat (values:VL006, VL0010,VL0612), quarter( values 1-4), GSA (a field including information for the Geographical Subarea in the following format: GSA20), Country in the following format (GRC)++}

\item{Sub.region}{The field of GSAs included in data using "".}

\item{Country}{The field of Country included in data using "".}

\item{FDays}{The field that contains information expressed in days at sea in a fine scale included the data table using "".}

\item{table_g_effort}{The data frame of the FDI Effort by country.csv. The above table is based on Table G in the data call (Fisheries Dependent Information (FDI) data)}

\item{gear}{The field in FPI table that contains the gear using "".Values: GNS, GTR, LLS}

\item{metier}{The field in FPI table that contains the metier using "".}

\item{fishing_tech}{The field in FPI table that contains the fishing technique using "".}

\item{target_assemblage}{The field in FPI table that contains the target assemblage using "".}

\item{year}{The field in FPI table that contains the year using "".}

\item{quarter}{The field in FPI table that contains the quarter using "".}

\item{vessel_length_cat}{The field in FPI table that contains the vessel length category using "". Values: VL006, VL0010, VL0612}

\item{LON}{Name of Longitude field in FPI table using "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FPI table "". The latitude should be in decimal degrees and in a geographical coordinate system WGS84}
}
\value{
Return of a data.frame including FPI and Fishing days by grid cell.
}
\description{
Combine Fishing effort estimations expressed in fishing days with Fishing Pressure Index (FPI- estimated by MCDA) with Fisheries Dependent Information (FDI) data (Table G in the data call) to perform comparisons by of the outcomes by GSA/gear/metier/quarter/fishing_tech/target_assemblage/year/country.
}
\examples{
library(tools4MCDA)

#load Data
data(FPIe)
data(ERS_poly)

# Define country
FPIe$Country<-"GRC"

# Estimate Fishing Days in a fine scale
result_q<- FDays_SSF(data=FPIe,Sub.region="Sub.region", Country="Country", FPIc="FPI", 
                    table_g_effort=table_g_effort,vessel_length_cat="vessel_length",
                    year="year", quarter="quarter",gear="gear_type",fishing_tech="fishing_tech",
                    target_assemblage="target_assemblage", metier="metier",LON="LON", LAT="LAT")

# Estimate Fishing Days by rectangle 0.5*0.5 decimal degrees
ERS_q<-Fdays_ERSlevel(data_FPI=result_q, ERS_poly=ERS_poly, LON="LON",LAT="LAT",parameter= "FDays")

result_q<- FDays_SSF(data=FPIe,Sub.region="Sub.region", Country="Country", FPIc="FPI", 
                    table_g_effort=table_g_effort,vessel_length_cat="vessel_length",
                    year="year", quarter="quarter",gear="gear_type",fishing_tech="fishing_tech",
                    target_assemblage="target_assemblage", metier="metier",LON="LON", LAT="LAT")

#check Days at sea result
check_Fdays<-check_FDays(data=result_q,Sub.region="Sub.region",Country="Country",FDays="FDays",
                        table_g_effort=table_g_effort,gear="gear_type",year="year",quarter="quarter",
                        vessel_length_cat="vessel_length",fishing_tech="fishing_tech", target_assemblage="target_assemblage",
                        metier="metier",LON="LON",LAT="LAT")
}
\author{
I. Maina
}
