% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_weight_rv030424_RDBFIS.R
\name{check_landings}
\alias{check_landings}
\title{Compare fishing effort in days at sea estimated using Fishing Pressure Index (from MCDA) as a proxy with table_A on landings estimations by GSA/gear/metier/quarter/fishing_tech/target_assemblage/year/country/species}
\usage{
check_landings(
  data,
  Sub.region,
  Country,
  FE,
  LPUE,
  species,
  table_a,
  gear,
  year,
  quarter,
  metier,
  fishing_tech,
  target_assemblage,
  vessel_length_cat,
  parameter = "totwghtlandg",
  LW,
  LON,
  LAT
)
}
\arguments{
\item{data}{A data frame with FE estimations (e.g. by MCDA and FDI) and species LPUE produced after running the tools4MCDA::landings_SSF function. The above data frame should contain the fields: LON(longitude in WGS84), LAT(latitude in WGS84), gear (values: GNS, GTR, LLS), year, vessel_length_cat , quarter(values 1-4), GSA (a field including information for the Geographical Subarea in the following format: GSA20, Country in the following format : GRC), FE (estimated fishing effort or the Fishing Pressure Index by MCDA), LPUE(estimated landings per unit effort)}

\item{Sub.region}{The field of GSAs included in data using "".}

\item{Country}{The field of Country included in data using "".}

\item{FE}{The field in the FPI table that contains the fishing effort or the Fishing Pressure Index by MCDA using ""}

\item{LPUE}{The field in the FPI table that contains the Landings Per Unit Effort using ""}

\item{species}{The field in the FPI table that contains the species using ""}

\item{table_a}{The data frame of the FDI "catches" (landing weight, landing value, discard weight) by country.csv. The above table is based on Table A in the data call (Fisheries Dependent Information (FDI) data and contains information about landing weight and value of landings)}

\item{gear}{The field in FPI table that contains the gear using "". Values: SSF, GNS, GTR, LLS}

\item{year}{The field in FPI table that contains the year using "".}

\item{quarter}{The field in FPI table that contains the quarter using ""}

\item{metier}{The field in FPI table that contains the metier using ""}

\item{fishing_tech}{The field in FPI table that contains the fishing technique using ""}

\item{target_assemblage}{The field in FPI table that contains the target assemblage using ""}

\item{vessel_length_cat}{The field in FPI table that contains the vessel length category using "". Values: VL006, VL0010, VL0612}

\item{parameter}{The field from table A to be used for performing spatial estimations. Values: "totwghtlandg", "totvallandg"}

\item{LW}{The field in the FPI table that contains the total weight of landings or value by species using "". This is estimated from tools4MCDA::landings_SSF (NOT REVISE)}

\item{LON}{Name of Longitude field in FPI table using "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FPI table "". The latitude should be in decimal degrees and in a geographical coordinate system WGS84}
}
\value{
Return of a data.frame including landing weight or value.
}
\description{
Use the produced information that combines Fishing effort estimations expressed in fishing days with Fishing Pressure Index (FPI- estimated by MCDA) to perform comparisons with Fisheries Dependent Information (FDI) data (Table A in the data call) by GSA/gear/metier/quarter/fishing_tech/target_assemblage/year/country/species
}
\examples{
library(tools4MCDA)
library(ggplot2)

data(FPIe_a)
data(table_a)

FPIe_a$Country<-"GRC"

#Estimate catch from FPI, table a (GSA, country, mandatory fields)
result_q_catch<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",parameter = "totwghtlandg",LON="LON",LAT="LAT")

result_q_value<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",parameter = "totvallandg",LON="LON",LAT="LAT")

#check Landing weight result
check_weight<-check_landings(data=result_q_catch,Sub.region="Sub.region", Country="Country", 
                          FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                          quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                          metier="metier",LON="LON",LAT="LAT",parameter = "totwghtlandg", LW="LW")

#check Landing value result
check_value<-check_landings(data=result_q_value,Sub.region="Sub.region", Country="Country", 
                           FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                           quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                         metier="metier",LON="LON",LAT="LAT",parameter = "totvallandg", LW="LW")


}
\author{
I. Maina
}
