% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl_nc_to_df.R
\name{chl_nc_to_df}
\alias{chl_nc_to_df}
\title{Convert monthly Chlorophyll-a netCDF dataset to a dataframe}
\usage{
chl_nc_to_df(file_path, lonmax = 26, lonmin = 18, latmax = 43, latmin = 33)
}
\arguments{
\item{file_path}{The file path for the netCDF monthly file for Aqua-MODIS Chlorophyll-a concentration. The netCDF file is in mapped-SMI format.}

\item{lonmax}{Maximum longitude value of the selected area (in decimal degrees).}

\item{lonmin}{Minimum longitude value of the selected area (in decimal degrees).}

\item{latmax}{Maximum latitude value of the selected area (in decimal degrees).}

\item{latmin}{Minimum latitude value of the selected area (in decimal degrees).}
}
\value{
Return a dataframe including longitude, latitude, chlorophyll-a value, year, month and quarter.
}
\description{
Using the monthly Aqua-MODIS Chlorophyll-a concentration netCDF datasets and convert to a dataframe. The dataframe can be produced for a certain area. The Clorophyll-a netCDF datasets are based on the level-3 products produced and archived by the NASA Ocean Biology Processing Group (NASA Ocean Biology Processing Group, 2024, Level-3 Ocean Color Data, NASA Ocean Biology Distributed Active Archive Center. https://oceancolor.gsfc.nasa.gov/l3) .
}
\examples{
library(tools4MCDA)

#Convert netCDF to dataframe for a certain geographical area
#Example for Greece
chl_df<-chl_nc_to_df(file_path="~.nc", lonmax=26, lonmin=18, latmax=43, latmin=33)
}
\author{
I. Maina
}
