% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tableH_22122025.R
\name{create_tableH2}
\alias{create_tableH2}
\title{Converts the estimated spatial landings' weight and value in the format of table H of the FDI data call (version 2 - detailed)}
\usage{
create_tableH2(
  data = ERS_gns_06_q_catch,
  country = "Country",
  year = "year",
  quarter = "quarter",
  vessel_length = "vessel_length",
  fishing_tech = "fishing_tech",
  gear_type = "gear_type",
  target_assemblage = "target_assemblage",
  mesh_size_range = "mesh_size_range",
  metier = "metier",
  metier_7 = "metier_7",
  supra_region = "MBS",
  sub_region = "Sub.region",
  eez_indicator = "NA",
  geo_indicator = "NK",
  specon_tech = "NK",
  deep = "NA",
  rectangle_type = "05*05",
  latitude = "csq_y",
  longitude = "csq_x",
  c_square = "NA",
  species = "species",
  totwghtlandg = "parameter",
  totvallandg = "parameter2",
  confidential = "N"
)
}
\arguments{
\item{data}{The dataframe derived in ERS level and includes information of landings' weight and value}

\item{country}{The name of the field in the dataframe containing the country}

\item{year}{The name of the field in the dataframe containing the year}

\item{quarter}{The name of the field in the dataframe containing the quarter}

\item{vessel_length}{The name of the field in the dataframe containing the vessel length}

\item{fishing_tech}{The name of the field in the dataframe containing the fishing technique}

\item{gear_type}{The name of the field in the dataframe containing the gear type}

\item{target_assemblage}{The name of the field in the dataframe containing the target assemblage}

\item{mesh_size_range}{The name of the field in the dataframe containing the mesh_size_range}

\item{metier}{The name of the field in the dataframe containing the metier}

\item{metier_7}{The name of the field in the dataframe containing the metier_7}

\item{supra_region}{Value for supra_region (default is "MBS")}

\item{sub_region}{The name of the field of the dataframe containing the sub region (GSA)}

\item{eez_indicator}{Value for eez_indicator}

\item{geo_indicator}{Value for geo_indicator}

\item{specon_tech}{Value for specon_tech}

\item{deep}{Value for deep}

\item{rectangle_type}{Value for rectangle_type}

\item{latitude}{The name of the field in the dataframe containing the latitude}

\item{longitude}{The name of the field in the dataframe containing the longitude}

\item{c_square}{Value for c_square}

\item{species}{The name of the field in the dataframe containing the species}

\item{totwghtlandg}{The name of the field in the dataframe containing the landing weight}

\item{totvallandg}{The name of the field in the dataframe containing the landing value}

\item{confidential}{Value for confidential}
}
\value{
Return a dataframe in the format of table H (spatial landings) of the FDI data call.
}
\description{
Using the predicted spatial landings' weight and value in a csquare of 0.5 * 0.5 decimal degrees cell size to export a dataframe in the format of spatial landings (Table H) of the FDI data call (version 2 - detailed).
}
\examples{
library(tools4MCDA)

#load Data
data(FPIe_a)
data(ERS_poly)

# Define country
FPIe_a$Country<-"GRC"

# Estimate landings weight in a fine scale
result_q_catch<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",metier_7="metier_7",mesh_size_range="mesh_size_range",parameter = "totwghtlandg",LON="LON",LAT="LAT")

result_q_value<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",metier_7="metier_7",mesh_size_range="mesh_size_range",parameter = "totvallandg",LON="LON",LAT="LAT")

# Estimate landings weight by rectangle 0.5*0.5 decimal degrees
ERS_q_catch<-catch_ERSlevel(data_FPI=result_q_catch, ERS_poly=ERS_poly, parameter= "LW")
ERS_q_value<-catch_ERSlevel(data_FPI=result_q_value, ERS_poly=ERS_poly, parameter= "LW")

library(dbplyr)
ERS_q_value<-ERS_q_value \%>\% 
 dplyr::rename(
   parameter2 = parameter
 )

ERS_q_catch_value<-cbind(ERS_q_catch,ERS_q_value[15])

#create table H for the data call
tableH_landings_SSF<-create_tableH(data=ERS_q_catch_value,country="Country",	year="year",	quarter="quarter",	vessel_length="vessel_length",
                                  fishing_tech="fishing_tech", gear_type="gear_type",	target_assemblage="target_assemblage",
                                  mesh_size_range="mesh_size_range",	metier="metier",	metier_7="metier_7",	supra_region="MBS",
                                  sub_region="Sub.region",	eez_indicator="NA",	geo_indicator="NK",	specon_tech="NK",
                                  deep="NA",	rectangle_type="05*05",	latitude="csq_y",	longitude="csq_x",	c_square="NA",species="species",
                                  totwghtlandg="parameter",totvallandg="parameter2",	confidential="N")
}
\author{
I. Maina
}
