% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tableI.R
\name{create_tableI}
\alias{create_tableI}
\title{Converts the estimated spatial fishing effort in the format of table I of the FDI data call}
\usage{
create_tableI(
  data = ERS_gns_06_q,
  country = "Country",
  year = "year",
  quarter = "quarter",
  vessel_length = "vessel_length",
  fishing_tech = "fishing_tech",
  gear_type = "gear_type",
  target_assemblage = "target_assemblage",
  mesh_size_range = "NK",
  metier = "metier",
  metier_7 = "NK",
  supra_region = "MBS",
  sub_region = "Sub.region",
  eez_indicator = "NA",
  geo_indicator = "NK",
  specon_tech = "NK",
  deep = "NA",
  rectangle_type = "05*05",
  latitude = "csq_y",
  longitude = "csq_x",
  c_square = "NA",
  totfishdays = "parameter",
  confidential = "N"
)
}
\arguments{
\item{data}{The dataframe derived in ERS level and includes information of effort}

\item{country}{The name of the field in the dataframe containing the country}

\item{year}{The name of the field in the dataframe containing the year}

\item{quarter}{The name of the field in the dataframe containing the quarter}

\item{vessel_length}{The name of the field in the dataframe containing the vessel length}

\item{fishing_tech}{The name of the field in the dataframe containing the fishing technique}

\item{gear_type}{The name of the field in the dataframe containing the gear type}

\item{target_assemblage}{The name of the field in the dataframe containing the target assemblage}

\item{mesh_size_range}{Value for mesh_size_range}

\item{metier}{The name of the field in the dataframe containing the metier}

\item{metier_7}{Value for metier_7}

\item{supra_region}{Value for supra_region (default is "MBS")}

\item{sub_region}{The name of the field in the dataframe containing the sub region (GSA)}

\item{eez_indicator}{Value for eez_indicator}

\item{geo_indicator}{Value for geo_indicator}

\item{specon_tech}{Value for specon_tech}

\item{deep}{Value for deep}

\item{rectangle_type}{Value for rectangle_type}

\item{latitude}{The name of the field in the dataframe containing the latitude}

\item{longitude}{The name of the field in the dataframe containing the longitude}

\item{c_square}{Value for c_square}

\item{totfishdays}{The name of the field in the dataframe containing the fishing days}

\item{confidential}{Value for confidential}

\item{species}{The name of the field in the dataframe containing the species}
}
\value{
Return a dataframe in the format of table I (spatial effort) of the FDI data call.
}
\description{
Using the predicted spatial fishing effort in a csquare of 0.5 * 0.5 decimal degrees cell size to export a dataframe in the format of spatial effort (Table I) of the FDI data call.
}
\examples{
library(tools4MCDA)

#load Data
data(FPIe)
data(ERS_poly)

# Define country
FPIe$Country<-"GRC"

# Estimate Fishing Days in a fine scale
result_q<- FDays_SSF(data=FPIe,Sub.region="Sub.region", Country="Country", FPIc="FPI", 
                    table_g_effort=table_g_effort,vessel_length_cat="vessel_length",
                    year="year", quarter="quarter",gear="gear_type",fishing_tech="fishing_tech",
                    target_assemblage="target_assemblage", metier="metier",LON="LON", LAT="LAT")

# Estimate Fishing Days by rectangle 0.5*0.5 decimal degrees
ERS_q<-Fdays_ERSlevel(data_FPI=result_q, ERS_poly=ERS_poly, LON="LON",LAT="LAT",parameter= "FDays")

#create table I in the data call format
tableI_seffort_SSF<-create_tableI(data=ERS_q,country="Country",	year="year",	quarter="quarter",	vessel_length="vessel_length",
                                 fishing_tech="fishing_tech", gear_type="gear_type",	target_assemblage="target_assemblage",
                                 mesh_size_range="NK",	metier="metier",	metier_7="NK",	supra_region="MBS",
                                 sub_region="Sub.region",	eez_indicator="NA",	geo_indicator="NK",	specon_tech="NK",
                                 deep="NA",	rectangle_type="05*05",	latitude="csq_y",	longitude="csq_x",	c_square="NA",	
                                 totfishdays="parameter",confidential="N")
}
\author{
I. Maina
}
