% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5NetcdfToCsv.R
\name{ERA5NetcdfToCsv}
\alias{ERA5NetcdfToCsv}
\title{Convert netcdf containing meteorological data from Copernicus products to a data.frame and store in a csv.}
\usage{
ERA5NetcdfToCsv(nc_data)
}
\arguments{
\item{nc_data}{The path that netcdf data are stored.}
}
\value{
Return of a data.frame.
}
\description{
Convert netcdf containing meteorological data to data.frame to be used in the function meteoCr. The data are based on ERA5 hourly data on single levels from 1940 to present from Copernicus Products (Spatial resolution: 0.25*0.25 decimal degrees).
Citation:
-Hersbach, H., Bell, B., Berrisford, P., Biavati, G., Horanyi, A., Munoz Sabater, J., Nicolas, J., Peubey, C., Radu, R., Rozum, I., Schepers, D., Simmons, A., Soci, C., Dee, D., Thepaut, J-N. (2023): ERA5 hourly data on single levels from 1940 to present. Copernicus Climate Change Service (C3S) Climate Data Store (CDS), DOI: 10.24381/cds.adbb2d47 (Accessed on DD-MMM-YYYY)
-Copernicus Climate Change Service, Climate Data Store, (2023): ERA5 hourly data on single levels from 1940 to present. Copernicus Climate Change Service (C3S) Climate Data Store (CDS), DOI: 10.24381/cds.adbb2d47 (Accessed on DD-MMM-YYYY)
}
\examples{


  # Example 
  library(ncdf4) # package for netcdf manipulation
  library(raster) # package for raster manipulation
  library(rgdal) # package for geospatial analysis
  library(ggplot2) 
  library(RNetCDF)

  meteo_data<-ERA5NetcdfToCsv(nc_data = ".ERA5data.nc")

}
\author{
I. Maina
}
