% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdays_ERSlevel.R
\name{Fdays_ERSlevel}
\alias{Fdays_ERSlevel}
\title{Convert the estimated fishing effort for SSF expressed in days at sea using Fishing Pressure Index as a proxy in a grid cell of 0.5*0.5 decimal degrees (csquare) in order to inform Table I.}
\usage{
Fdays_ERSlevel(
  data_FPI,
  ERS_poly,
  LON = "LON",
  LAT = "LAT",
  parameter = "FDays"
)
}
\arguments{
\item{data_FPI}{A data frame with Fishing effort estimations after running the function FDays_SSF. The above data frame should contain the fields: Fishing Effort, LON(longitude in WGS84), LAT(latitude in WGS84), gear:(values: GNS, GTR, LLS), year, vessel_length_cat (values: "VL0006","VL0010","VL0612",), quarter (values: 1-4), GSA (a field including information for the Geographical Subarea in the following format: GSA20, Country in the following format : GRC)}

\item{ERS_poly}{A polygon shapefile with ERS rectangle (csquare of 0.5*0.5) in an 'sf' format}

\item{LON}{Name of Longitude field in FPI table using "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FPI table "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{FDays}{Name of the fishing days in the data table in "".}
}
\value{
Return of a data.frame including Fishing days by grid cell of 0.5*0.5 decimal degrees to be used for the creation of table I.
}
\description{
Convert in csquare (0.5 * 0.5 dd) the combined Fishing Pressure Index (FPI- estimated by MCDA) with Fisheries Dependent Information (FDI) data (Table G in the data call) to perform spatial estimations of fishing effort (days at sea) and inform the Table I.
}
\examples{
library(tools4MCDA)
#load Data
data(FPIe)
data(ERS_poly)

# Define country
FPIe$Country<-"GRC"

# Estimate Fishing Days in a fine scale
result_q<- FDays_SSF(data=FPIe,Sub.region="Sub.region", Country="Country", FPIc="FPI", 
                    table_g_effort=table_g_effort,vessel_length_cat="vessel_length",
                    year="year", quarter="quarter",gear="gear_type",fishing_tech="fishing_tech",
                    target_assemblage="target_assemblage", metier="metier",LON="LON", LAT="LAT")

# Estimate Fishing Days by rectangle 0.5*0.5 decimal degrees
ERS_q<-Fdays_ERSlevel(data_FPI=result_q, ERS_poly=ERS_poly, LON="LON",LAT="LAT",parameter= "FDays")
}
\author{
I. Maina
}
