% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dsea_SSF_rv22122025_4RDBFIS.R
\name{FDays_SSF2}
\alias{FDays_SSF2}
\title{Estimate fishing effort for SSF expressed in days at sea using Fishing Pressure Index as a proxy (version 2 - detailed).}
\usage{
FDays_SSF2(
  data,
  Sub.region,
  Country,
  FPIc,
  table_g_effort,
  gear,
  metier,
  metier_7,
  mesh_size_range,
  year,
  quarter,
  vessel_length_cat,
  fishing_tech,
  target_assemblage,
  LON,
  LAT
)
}
\arguments{
\item{data}{A data frame with FPI estimations (MCDA). The above data frame should contain the fields: LON(longitude in WGS84), LAT(latitude in WGS84), gear(values: GNS, GTR, LLS), year, vessel_length_cat (values: "VL0006","VL0612"), quarter(values: 1-4), GSA (a field including information for the Geographical Subarea in the following format: GSA20, Country in an Alpha 3 code format : GRC)}

\item{Sub.region}{The field of GSAs included in data using "".}

\item{Country}{The field of Country included in data using "".}

\item{FPIc}{The field of Fishing pressure Index estimated by MCDA and included in data using "".}

\item{table_g_effort}{a dataframe similar to Table G (as in the FDI data call)}

\item{gear}{The field in FPI table that contains the gear using "".Values: SSF, GNS, GTR, LLS}

\item{metier}{The field in FPI table that contains the metier using "". Values should be the same as in table G}

\item{metier_7}{The field in FPI table that contains the metier_7 using "". Values should be the same as in table G}

\item{mesh_size_range}{The field in FPI table that contains the mesh_size_range using "". Values should be the same as in table G}

\item{year}{The field in FPI table that contains the year using "".}

\item{quarter}{The field in FPI table that contains the quarter using "". Values: 1-4.}

\item{vessel_length_cat}{The field in FPI table that contains the vessel length category using "". Values: VL006, VL0612}

\item{fishing_tech}{The field in FPI table that contains the fishing_tech using "". Values should be the same as in table G}

\item{target_assemblage}{The field in FPI table that contains the target assemblage using "". Values should be the same as in table G}

\item{LON}{Name of Longitude field in FPI table using "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FPI table "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}
}
\value{
Return of a data.frame including FPI and Fishing days by grid cell.
}
\description{
Combine Fishing Pressure Index (FPI- estimated by MCDA) with Fisheries Dependent Information (FDI) data (Table G in the data call) to perform spatial estimations of fishing effort (days at sea) (version 2 - detailed).
}
\examples{
library(tools4MCDA)

#load Data
data(FPIe)
data(ERS_poly)

# Define country
FPIe$Country<-"GRC"

# Estimate Fishing Days in a fine scale
result_q<- FDays_SSF(data=FPIe,Sub.region="Sub.region", Country="Country", FPIc="FPI", 
                    table_g_effort=table_g_effort,vessel_length_cat="vessel_length",
                    year="year", quarter="quarter",gear="gear_type",fishing_tech="fishing_tech",
                    target_assemblage="target_assemblage", metier="metier",metier_7="metier_7",mesh_size_range="mesh_size_range",
                    LON="LON", LAT="LAT")
}
\author{
I. Maina
}
