% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCR.R
\name{getCR}
\alias{getCR}
\title{Analytic Hierarchy Process (AHP)}
\usage{
getCR(pairwise_matrix)
}
\arguments{
\item{pairwise_matrix}{The nxn pair-wise comparison matrix.}
}
\value{
\item{cr}{The consistency ratio.}

\item{pairwise_matrix_adj}{ The adjusted pairwise matrix with acceptable
consistency (C.R. < 0.1).  }

\item{delta}{ Measure the maximum absolute difference of elements between
the pairwise matrix and adjusted pairwise matrix. It should be <2.  }

\item{sig}{ A metric that measures the standard deviation of the difference
of elements between the pairwise matrix and adjusted pairwise matrix. It
should be ~1.  }
}
\description{
Compute the consistency ratio of the initial pair-wise comparison matrix
}
\examples{
library(FuzzyAHP)
library(tools4MCDA)
matrix_values <- c(1,3,4,5,7,
                  1/3,1,2,2,4,
                  1/4,1/2,1,2,3,
                  1/5,1/2,1/2,1,3,
                  1/7,1/4,1/3,1/3,1)
matrix_values = matrix(matrix_values,nrow = 5,ncol = 5, byrow = TRUE)
ahp_outputs = getCR(matrix_values)
ahp_outputs
}
\references{
Kavadas, S., I. Maina, D. Damalas, I. Dokos, M. Pantazi, and V.
Vassilopoulou (2015). Multi-Criteria Decision Analysis as a tool to extract
fishing footprints and estimate fishing pressure: application to small scale
coastal fisheries and implications for management in the context of the
Maritime Spatial Planning Directive. Mediterranean Marine Science
16:294-304. \url{http://dx.doi.org/10.12681/mms.1087}

Saaty, T.L. (2001). Decision Making for Leaders: The Analytic Hierarchy
Process for Decisions in a Complex World, New Edition 2001 (3 Revised).
Pittsburgh, PA: RWS Publications, ISBN 978-0962031786.

Xu, Z. (2004). A practical method for improving consistency of judgment
matrix in the ahp. Journal of Systems Science and Complexity, 17(2).
}
\author{
D. Politikos
}
