% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_extent.R
\name{incr_extent}
\alias{incr_extent}
\title{Increase the extent of a dataframe}
\usage{
incr_extent(
  df = chl_y_q,
  tolerance = 2,
  LON = "x",
  LAT = "y",
  year = "year",
  quarter = "quarter"
)
}
\arguments{
\item{df}{a dataframe with the following fields: longitude, latitude, quarter (values: 1-4), year.}

\item{tolerance}{a number to increase the extent}

\item{LON}{the name of longitude field using ""}

\item{LAT}{the name of latitude field using ""}

\item{year}{the name of year field using ""}

\item{quarter}{the name of quarter field using ""}
}
\value{
Returns a dataframe
}
\description{
Add longitude and latitude to increase the extent of a dataframe based on a tolerance value. The dataframe should include the following fields: longitude, latitude, quarter (values: 1-4), year.
}
\examples{
library(tools4MCDA)
data(chl_y_q)
chl_increase_extent<-incr_extent(df = chl_y_q, tolerance=2, LON="x", LAT="y", year="year", quarter="quarter")
}
\author{
I. Maina
}
