% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinFPI_GSA.R
\name{joinFPI_GSA}
\alias{joinFPI_GSA}
\title{Combine Fishing Pressure Index (FPI- estimated by MCDA) with GSA polygons to estimate the Geographical Sub-Area.}
\usage{
joinFPI_GSA(data_FPI, GSA_poly, field_GSA = "SMU_CODE", LON, LAT)
}
\arguments{
\item{data_FPI}{A data frame with FPI estimations. The above data frame should contain the fields: LON(longitude in WGS84), LAT(latitude in WGS84), gear (values- SSF, GNS, GTR, LLS), year, vessel_length_cat , quarter( values 1-4, or 0 for annual estimations)}

\item{GSA_poly}{The sf polygon of GSAs.}

\item{field_GSA}{The field that contains numerical information for the GSA using "".}

\item{LON}{Name of Longitude field in FDI table. The longitude should be in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FDI table. The longitude should be in a geographical coordinate system WGS84}
}
\value{
Return of a data.frame of FPI with information by GSA.
}
\description{
Combine Fishing Pressure Index (FPI- estimated by MCDA) with GSA polygons to estimate the Geographical Sub-Area.
}
\examples{
library(tools4MCDA)
library(sf)

data("dfcut2")
data("GSA_poly")

dfcut2<-joinFPI_GSA(data_FPI=dfcut2,GSA_poly = GSA_poly,field_GSA="SMU_CODE",LON="x",LAT="y")
}
\author{
I. Maina
}
