% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_SSF_030424_RDBFIS.R
\name{landings_SSF}
\alias{landings_SSF}
\title{Estimate landing weight or value using Fishing effort and landings per unit effort as a proxy}
\usage{
landings_SSF(
  data,
  Sub.region,
  Country,
  FE,
  LPUE,
  species,
  table_a,
  gear,
  year,
  quarter,
  metier,
  fishing_tech,
  target_assemblage,
  vessel_length_cat,
  parameter = "totwghtlandg",
  LON,
  LAT
)
}
\arguments{
\item{data}{A data frame with FE estimations (e.g. by MCDA and FDI) and species LPUE. The above data frame should contain the fields: LON(longitude in WGS84), LAT(latitude in WGS84), gear (values- SSF, GNS, GTR, LLS), year, vessel_length_cat , quarter( values 1-4, or 0 for annual estimations), GSA (a field including information for the Geographical Subarea in the following format: GSA20, Country in the following format : Greece), FE (estimated fishing effort), LPUE(estimated landings per unit effort)}

\item{Sub.region}{The field of GSAs included in data using "".}

\item{Country}{The field of Country included in data using "".}

\item{FE}{the field in the FPI table that contains the fishing effort or the Fishing Pressure Index estimated from MCDA using ""}

\item{LPUE}{the field in the FPI table that contains the landings per unit effort by species using "". The values should be normalized in a scale: 0-1}

\item{species}{the field in the FPI table that contains the species (in 3 alpha code) using ""}

\item{table_a}{The data frame of the FDI "catches" (landing weight, landing value, discard weight) by country.csv. The above table is based on Table A in the data call (Fisheries Dependent Information (FDI) data and contains information about landing weight and value of landings)}

\item{gear}{The field in FPI table that contains the gear using "".Values: GNS, GTR, LLS}

\item{year}{The field in FPI table that contains the year using "".}

\item{quarter}{The field in FPI table that contains the quarter using "".}

\item{metier}{The field in FPI table that contains the metier using "".}

\item{fishing_tech}{The field in FPI table that contains the fishing technique using "".}

\item{target_assemblage}{The field in FPI table that contains the target assemblage using "".}

\item{vessel_length_cat}{The field in FPI table that contains the vessel length category using "". Values: NK, VL006, VL0010, VL0612, VL1012, VL1218, VL1824, VL2440, VL40XX}

\item{parameter}{The field from table a to be used for performing spatial estimations. Values: "totwghtlandg", "totvallandg"}

\item{LON}{Name of Longitude field in FPI table using "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}

\item{LAT}{Name of Latitude field in FPI table "". The longitude should be in decimal degrees and in a geographical coordinate system WGS84}
}
\value{
Return of a data.frame including FPI and Fishing days by grid cell.
}
\description{
Combine Fishing Pressure Index (FPI- estimated by MCDA) with Fisheries Dependent Information (FDI) data (Table A in the data call) to perform spatial estimations of fishing effort (days at sea). Subsequently, this information is combined with landings per unit effort that can be estimated through predictive models. Finally all the above information is combined with table A from FDI data call to estimate landing weight and landing value in a fine spatial scale (e.g. ERS grid). The function can be used for discards as well if this information is provided instead of LPUE.
}
\examples{
library(tools4MCDA)
library(ggplot2)

data(FPIe_a)
data(table_a)

FPIe_a$Country<-"GRC"

#Estimate catch from FPI, table a (GSA, country, mandatory fields)
result_q_catch<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",parameter = "totwghtlandg",LON="LON",LAT="LAT")


result_q_value<- landings_SSF(data=FPIe_a,Sub.region="Sub.region", Country="Country", 
                             FE="FPI", LPUE="LPUE", species="species", table_a=table_a,gear="gear_type",year="year",
                             quarter="quarter", vessel_length_cat="vessel_length",fishing_tech ="fishing_tech",target_assemblage="target_assemblage",
                             metier="metier",parameter = "totvallandg",LON="LON",LAT="LAT")
}
\author{
I. Maina
}
