% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_year.R
\name{maps_year}
\alias{maps_year}
\title{Mapping by year and fishing gear type}
\usage{
maps_year(
  data_df,
  years = ye,
  parameter = "",
  xlim = c(19, 24),
  ylim = c(36, 40),
  LON = "longitude",
  LAT = "latitude",
  ylab = "kg/km2",
  PlotTitle = "HKE",
  poly = poly,
  r = r
)
}
\arguments{
\item{data_df}{The dataframe to be plotted}

\item{years}{The field of years in the dataframe in ""}

\item{parameter}{The field in the dataframe that will be plotted. The field should be in "". For example "totfishdays".}

\item{xlim}{A numeric object with the minimum and maximum longitude values to be plotted. For example c(19,24)}

\item{ylim}{A numeric object with the minimum and maximum latitude values to be plotted. For example c(36,40)}

\item{LON}{The field of the dataframe including longitude values.The field should be in "".}

\item{LAT}{The field of the dataframe including latitude values.The field should be in "".}

\item{ylab}{The label in the legend of the plot in "".}

\item{PlotTitle}{The title of the plot in "".}

\item{poly}{A coastline polygon}

\item{r}{A raster object}
}
\value{
Return maps by year, quarter etc.
}
\description{
Create plots presenting the spatial distribution of fishing effort, landings weight or value by year, fishing gear type, metier etc.
}
\examples{
library(tools4MCDA)
library(ggplot2)
library(raster)
library(rasterVis)

data("sample_FPIq")
data("dfcut2")
poly<- map_data("world")

#DEFINE COUNTRY, GSA, year_start, year_end
country<-"GRC"
GSA<-20

#crop the area and create a raster
dfcut2$lon<-as.numeric(dfcut2$x)
dfcut2$lat<-as.numeric(dfcut2$y)
dfcut2<-joinFPI_GSA(data_FPI=dfcut2,GSA_poly = GSA_poly,field_GSA="SMU_CODE",LON="x",LAT="y")
dfcut2<-dfcut2[dfcut2$SMU_CODE \%in\% GSA,]
r<-rasterFromXYZ(dfcut2[, c('lon', 'lat')])
crs(r) <- CRS(paste('+init=EPSG:4326'))

#preparing outcomes from MCDA
FPI_all<-unite(sample_FPIq, col='key', c('year','vessel_length','gear_type'), sep='_', remove=FALSE)

#Select gear type and vessel length category
p<-"2019_VL0006_GNS"
result_q_4p<-FPI_all[FPI_all$key==p,]
result_q_4p<-result_q_4p[result_q_4p$quarter==1,]

maps_year(data_df=result_q_4p, years = 2019, parameter = "FPI", xlim = c(19,24), ylim = c(36, 40), LON = "LON", LAT = "LAT", 
         ylab = "proxy of FE", PlotTitle = "FPI 2019", poly = poly, r = r) 
}
\author{
I. Maina
}
