% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_year_quarter.R
\name{maps_year_quarter}
\alias{maps_year_quarter}
\title{Mapping by quarter, year and fishing gear type}
\usage{
maps_year_quarter(
  data_df,
  quarter = qu,
  parameter = "",
  xlim = c(19, 24),
  ylim = c(36, 40),
  LON = "longitude",
  LAT = "latitude",
  ylab = "kg/km2",
  PlotTitle = "HKE",
  poly = poly,
  r = r
)
}
\arguments{
\item{quarter}{A numeric object including the quarters that data are available. For example c(1,2,3,4) .}

\item{parameter}{The field in the dataframe that will be plotted. The field should be in "". For example "totfishdays".}

\item{xlim}{A numeric object with the minimum and maximum longitude values to be plotted. For example c(19,24)}

\item{ylim}{A numeric object with the minimum and maximum latitude values to be plotted. For example c(36,40)}

\item{LON}{The field of the dataframe including longitude values.The field should be in "".}

\item{LAT}{The field of the dataframe including latitude values.The field should be in "".}

\item{ylab}{The label in the legend of the plot in "".}

\item{PlotTitle}{The title of the plot in "".}

\item{poly}{A raster object}

\item{data}{The dataframe to be plotted}
}
\value{
Return maps by year, quarter etc.
}
\description{
Create plots presenting the spatial distribution of fishing effort, landings weight or value by year, quarter, fishing gear type, metier etc.
}
\examples{
library(tools4MCDA)
library(ggplot2)
library(raster)
library(rasterVis)

data("sample_FPIq")
data("dfcut2")
poly<- map_data("world")

#DEFINE COUNTRY, GSA, year_start, year_end
country<-"GRC"
GSA<-20

#crop the area and create a raster
dfcut2$lon<-as.numeric(dfcut2$x)
dfcut2$lat<-as.numeric(dfcut2$y)
dfcut2<-joinFPI_GSA(data_FPI=dfcut2,GSA_poly = GSA_poly,field_GSA="SMU_CODE",LON="x",LAT="y")
dfcut2<-dfcut2[dfcut2$SMU_CODE \%in\% GSA,]
r<-rasterFromXYZ(dfcut2[, c('lon', 'lat')])
crs(r) <- CRS(paste('+init=EPSG:4326'))

#preparing outcomes from MCDA
FPI_all<-unite(sample_FPIq, col='key', c('year','vessel_length','gear_type'), sep='_', remove=FALSE)

#Select gear type and vessel length category
p<-"2019_VL0006_GNS"
result_q_4p<-FPI_all[FPI_all$key==p,]

#estimate quarter  
qu<-order(as.numeric(unique(result_q_4p$quarter)), decreasing = TRUE)

#Create maps
maps_year_quarter(data_df=result_q_4p,quarter=qu, parameter="FPI",xlim=c(min(result_q_4p$LON),max(result_q_4p$LON)),
                 ylim=c(min(result_q_4p$LAT),max(result_q_4p$LAT)), LON="LON", LAT="LAT",ylab="FPI",
                 PlotTitle=paste(unique(result_q_4p$key)), 
                 poly=poly, r=r)
}
\author{
I. Maina
}
