% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteoCr.R
\name{meteoCr}
\alias{meteoCr}
\title{Create meteorological criterion}
\usage{
meteoCr(
  data,
  time_start = 21,
  time_end = 4,
  temp_step = "quarter",
  diff_temp = 4
)
}
\arguments{
\item{data}{The path that meteorological data or a data frame generated by ERA5NetcdfToCsv or MeteoNetcdfToCsv are stored .}

\item{time_start}{Time (expressed in hours) of a fishing trip starts.}

\item{time_end}{Time (expressed in hours) of a fishing trip ends.}

\item{temp_step}{Temporal step of the outcome. Values: "quarter" (default), "month", "year"}

\item{diff_temp}{Parameter for estimating the extreme temperature Values. The default is 4 that means the extreme temperature can be 4 Co less that the max value.}
}
\value{
Return of a data.frame.
}
\description{
Score a netcdf spatiotemporal dataset based on meteorological criteria.
}
\examples{


  # Example 1
  library(tools4MCDA) # package for raster manipulation
 
  meteo_data<-ERA5NetcdfToCsv(nc_data = ".ERA5data.nc")
  meteo2017_q<-meteoCr(data = meteo_data, time_start=21, time_end=4, temp_step="quarter", diff_temp=4)
}
\author{
I. Maina
}
