% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeteoNetcdfToCsv.R
\name{MeteoNetcdfToCsv}
\alias{MeteoNetcdfToCsv}
\title{Convert netcdf containing meteorological data to csv.}
\usage{
MeteoNetcdfToCsv(nc_data)
}
\arguments{
\item{nc_data}{The path that netcdf data are stored.}
}
\value{
Return of a data.frame.
}
\description{
Convert netcdf containing meteorological data to csv. Outcomes can be used in the function meteoCr.
}
\examples{


  # Example 
  library(ncdf4) # package for netcdf manipulation
  library(raster) # package for raster manipulation
  library(ggplot2) 

  meteo2018_q<-MeteoNetcdfToCsv(nc_data = ".data.nc")
}
\author{
I. Maina
}
